Fixed issue causing tests to fail

+++ backend/api/core/views/services.py
@@ -50,27 +50,23 @@ def process_post_form(request, form):
             post.save()
             update_interests_and_hashtags(post)
             trends.analyze(post) # TODO this is potentially very expensive
-            
-            poll = Poll.objects.create(post=post, question="Your poll question here")  # Create a Poll associated with the Post
-            
-            for i in range(1, poll_choices + 1):
-                choice_text = form.cleaned_data.get(f'choice_{i}')
-                if (choice_text != ''):
-                    # Pass text to NLP model
-                    choice_result = classify_text(choice_text)
-                    
-                    if choice_result["prediction"][0] in [1, 0]:  # Offensive or hate speech in choice text
-                        message = 'One of the poll choices contains offensive language and is not allowed on our platform.' if choice_result["prediction"][0] == 1 else 'One of the poll choices contains hateful language and is not allowed on our platform.'
-                        messages.error(request, message)
-                        post.delete()  # Delete the post if any poll choice is inappropriate
-                        return None
-                    
-                    elif choice_result["prediction"][0] == 2:
-                        # Create a poll with the given choices
-                        PollChoice.objects.create(poll=poll, choice_text=choice_text)
-            
-            update_interests_and_hashtags(post)
-            trends.analyze(post) # TODO this is potentially very expensive
+            if (poll_choices): 
+                poll = Poll.objects.create(post=post) 
+                for i in range(1, poll_choices + 1):
+                    choice_text = form.cleaned_data.get(f'choice_{i}')
+                    if (choice_text != ''):
+                        # Pass text to NLP model
+                        choice_result = classify_text(choice_text)
+                        
+                        if choice_result["prediction"][0] in [1, 0]:  # Offensive or hate speech in choice text
+                            message = 'One of the poll choices contains offensive language and is not allowed on our platform.' if choice_result["prediction"][0] == 1 else 'One of the poll choices contains hateful language and is not allowed on our platform.'
+                            messages.error(request, message)
+                            post.delete()  # Delete the post if any poll choice is inappropriate
+                            return None
+                        
+                        elif choice_result["prediction"][0] == 2:
+                            # Create a poll with the given choices
+                            PollChoice.objects.create(poll=poll, choice_text=choice_text)
             return post    
     return None
